update main.h, options.h and main.cpp
