+++ include/main.h
@@ -64,6 +64,12 @@
 using std::vector;
 using std::string;
 
+// Our namespace
+#include "hyper-base.hpp"
+#include "hyper-components.hpp"
+#include "hyper-chassis.hpp"
+#include "hyper-auton.hpp"
+
 /**
  * If you find doing pros::Motor() to be tedious and you'd prefer just to do
  * Motor, you can use the namespace with the following commented out line.
